+++ src/main/java/com/codeup/adlister/dao/Planets.java
@@ -10,8 +10,8 @@ public interface Planets {
     List<Planet> all();
     // insert a new ad and return the new ad's id
     Long insert(Planet planet);
-    Planet findPlanetById(long id);
-    List<Planet> usersPlanets(long id);
+    Planet findPlanetById(int id);
+    List<Planet> usersPlanets(int id);
     int updatePlanet(Planet planet);
-    int deletePlanet(long id);
+    int deletePlanet(int id);
 }
